+++ rubyllm_chat_app/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [0.8.24] - 2026-07-15
-🔵 🖼️ Fixed regenerate image button by purging old image and sleeping 1s explicitly.
+
+* Fix bug #44: Add `status: :see_other` to Turbo Stream redirects in `ChatsController` 🐛
 
 ## v0.8.23 - on 20260713